:sparkler: refactor sonarqube and add JSON parsing for api export
